// generated/src/app/api-models/case-mgmt-auth-send-rule.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CaseMgmtAuthSendRule {
  caseMgmtAuthSendRuleSeq: number;
  caseMgmtAuthSendRuleDet: string;
  valueFrom: string;
  valueThru: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}